remove upload handler
